auth-service/requirements.txt
fastapi
uvicorn
python-dotenv
databases
sqlalchemy
asyncpg
python-jose

httpx
pydantic[email]
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
